added on screen kbd changed canvas structure
