+++ package.json
@@ -20,6 +20,7 @@
     "bootstrap": "^5.3.8",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
+    "react-simple-keyboard": "^3.8.122",
     "twgl.js": "^7.0.0",
     "zustand": "^5.0.8"
   },
