+++ src/_custom.scss
@@ -16,11 +16,6 @@
   background-color: $blue-800;
   border-color: $blue-300;
   letter-spacing: 0.2rem;
-  &_start {
-    width: 10rem;
-    height: 10rem;
-    border-radius: 100%;
-  }
 
   &_rounded {
     background-color: $gray-800;
@@ -41,13 +36,15 @@
 
 .input {
   &_rounded {
-    font-size: 1rem;
+    font-size: 2rem;
     padding: 1rem;
     outline: none !important;
     box-shadow: none !important;
     background-color: $gray-800;
     border: 1px solid $gray-300;
     border-radius: 1rem;
+    min-height: 15rem;
+    text-align: start;
   }
 }
 
@@ -74,4 +71,36 @@
   background-color: #000000cc;
   width: 100%;
   height: 100%;
+}
+
+// override keyboard styles
+
+.simple-keyboard.hg-theme-default.myTheme {
+  touch-action: manipulation;
+  border-radius: 10px;
+  background-color: var(--bs-body-bg);
+
+  margin: -3rem;
+  width: calc(100% + 6rem);
+  height: 25vh;
+
+  .hg-rows {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+  }
+
+  .hg-row {
+    flex-grow: 1;
+    display: flex;
+    height: 100%;
+  }
+  .hg-button {
+    flex-grow: 1;
+    height: 100%;
+
+    background-color: #333333;
+    color: white;
+  }
 }
