+++ src/constants.tsx
@@ -1,11 +1,17 @@
 export const START_SCREEN = 0;
 export const SEQUENCE = [
-  "Start",
+  // "Start",
   "Message",
-  "VisualVibe",
-  // "Mood",
+  "Style",
+  "Mood",
   "Done",
 ] as const;
 
 export const CIRCLE_RADIUS = 40;
 export const POLLING_TIME = 1000;
+export const KEYBOARD_LAYOUT = [
+  "Q W E R T Y U I O P",
+  "A S D F G H J K L",
+  "Z X C V B N M {bksp}",
+  "{space}",
+];
