+++ src/pages/components/Text.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from "react";
 
 export function LargeText({ children }: { children: ReactNode }) {
-  return <p className="text-start mfs_1 text-uppercase">{children}</p>;
+  return <p className="text-start mfs_1">{children}</p>;
 }
